+++ Changelog.md
@@ -5,6 +5,8 @@ Changelog [format](http://keepachangelog.com/en/1.0.0/)
 ## Unreleased
 
 ### Changed
+* GODT-388 support for both bridge and import/export credentials by package users
+* GODT-387 store factory to make store optional
 * GODT-386 renamed bridge to general users and keep bridge only for bridge stuff
 * GODT-308 better user error message when request is canceled
 * GODT-312 validate recipient emails in send before asking for their public keys